added fractional shares

+++ app/core/calc_v2/portfolio_allocator/models.py
@@ -65,7 +65,7 @@ class Allocation(BaseModel):
     """Single ticker allocation with weight and share count."""
     ticker: str
     weight: float
-    num_shares: int
+    num_shares: float
 
 
 class PortfolioPerformance(BaseModel):
